ITE412_SIA2_AbSentinels_AniPalengke_Project/src/backend/requirements.txt
Django==4.2.14
djangorestframework==3.15.2
PyJWT==2.9.0
